finished database migration; switched to HttpClient; fixed issue with engraving modal;
