+++ LostArkBot/LostArkBot.csproj
@@ -8,21 +8,21 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Discord.Net" Version="3.7.2"/>
-        <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.8"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.8"/>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.8"/>
+        <PackageReference Include="Discord.Net" Version="3.7.2" />
+        <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="6.0.8" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.8" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.8" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.0.8">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0"/>
-        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0"/>
-        <PackageReference Include="MySql.Data" Version="8.0.30"/>
-        <PackageReference Include="MySql.EntityFrameworkCore" Version="6.0.4"/>
-        <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="6.0.2"/>
-        <PackageReference Include="PuppeteerSharp" Version="7.1.0"/>
-        <PackageReference Include="Quartz" Version="3.4.0"/>
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.0.0" />
+        <PackageReference Include="MySql.Data" Version="8.0.30" />
+        <PackageReference Include="MySql.EntityFrameworkCore" Version="6.0.4" />
+        <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="6.0.2" />
+        <PackageReference Include="PuppeteerSharp" Version="7.1.0" />
+        <PackageReference Include="Quartz" Version="3.4.0" />
     </ItemGroup>
 
     <ItemGroup>
